+++ src/SCBE_Coff.cpp
@@ -653,7 +653,7 @@ void SCBE_Coff::computeUnwind(const VectorNative<CPURegister>& unwindRegs,
     for (int32_t i = static_cast<int32_t>(unwindRegs.size()) - 1; i >= 0; i--)
     {
         uint16_t unwind0 = 0;
-        unwind0          = unwindRegs[i] << 12;
+        unwind0          = static_cast<uint16_t>(unwindRegs[i] << 12);
         unwind0 |= UWOP_PUSH_NO_VOL << 8;
         unwind0 |= static_cast<uint8_t>(unwindOffsetRegs[i]);
         unwind.push_back(unwind0);
